uboot-2025: restructure failsafe code --- separate functions to improve maintainability
