+++ uboot-mtk-20250711/failsafe/Makefile
@@ -9,5 +9,8 @@ obj-y += fsdata/
 endif
 
 obj-y += fs.o failsafe.o
+obj-$(CONFIG_WEBUI_FAILSAFE_CONSOLE) += failsafe_console.o
+obj-$(CONFIG_WEBUI_FAILSAFE_ENV) += failsafe_env.o
+obj-$(CONFIG_WEBUI_FAILSAFE_BACKUP) += failsafe_backup_flash.o
 
 endif
